Revert "Revert "Changed all mentions in TransferMoney from decimal to double.""
This reverts commit 16908ae385ec7eabc47a41912d991550900ce97d.

+++ CurrencyConverter.cs
@@ -8,21 +8,21 @@ namespace The_Bank
 {
     public static class CurrencyConverter
 {
-    public static decimal Convert(string sourceCurrency, string destinationCurrency, decimal amount)
+    public static double Convert(string sourceCurrency, string destinationCurrency, double amount)
     {
 
-            Dictionary<string, decimal> exchangeRates = new Dictionary<string, decimal>
+            Dictionary<string, double> exchangeRates = new Dictionary<string, double>
             {
                 { "SEK", 1 },
-                { "USD", 0.094M },  //M betyder att värdet är literal (Typ bokstavligen, utan detta blir det errors. THANKS GOOGLE
-                { "EUR", 0.087M },
-                { "GBP", 0.076M},
-                { "CHF", 0.084M},
-                { "CAD", 0.13M },
-                { "ZWD", 225.18M},
+                { "USD", 0.094 },  //M betyder att värdet är literal (Typ bokstavligen, utan detta blir det errors. THANKS GOOGLE
+                { "EUR", 0.087 },
+                { "GBP", 0.076},
+                { "CHF", 0.084},
+                { "CAD", 0.13 },
+                { "ZWD", 225.18},
             };
 
-            decimal convertedAmount = amount * exchangeRates[destinationCurrency];
+            double convertedAmount = amount * exchangeRates[destinationCurrency];
             return convertedAmount;
